fix loss mask

+++ utils/loss_rtdetr.py
@@ -415,7 +415,6 @@ class DETRLoss(nn.Module):
         return {k: v.squeeze() for k, v in loss.items()}
 
     def _get_loss_mask(self, masks, gt_mask, match_indices, gt_groups, postfix=''):
-        # masks: [b, query, h, w], gt_mask: list[[n, H, W]]
         name_mask = f'loss_mask{postfix}'
         name_dice = f'loss_dice{postfix}'
 
@@ -429,6 +428,7 @@ class DETRLoss(nn.Module):
         num_gts = sum(gt_groups)
         src_masks, target_masks = self._get_assigned_bboxes(masks, gt_mask, match_indices, gt_groups)
         src_masks = F.interpolate(src_masks.unsqueeze(0), size=target_masks.shape[-2:], mode='bilinear')[0]
+        target_masks = target_masks.float()  # Chuyển target_masks sang float32
         loss[name_mask] = self.loss_gain['mask'] * sigmoid_focal_loss(src_masks, target_masks, num_gts) / len(target_masks)
         loss[name_dice] = self.loss_gain['dice'] * self._dice_loss(src_masks, target_masks, num_gts) / len(target_masks)
         return loss
@@ -579,7 +579,7 @@ class DETRLoss(nn.Module):
                     for t, (_, j), k in zip(gt_bboxes, match_indices, range(len(gt_bboxes)))
                     if len(t) > 0  # Bỏ qua tensor rỗng
                 ]
-            )
+            ).float()  # Chuyển sang float32 cho mask
 
         return pred_assigned, gt_assigned
